backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
aiomysql==0.2.*
alembic==1.14.*
pydantic-settings==2.7.*
pyjwt==2.10.*
bcrypt==4.2.*
python-multipart==0.0.*
email-validator==2.2.*
markdown==3.7.*
feedgen==1.0.*
bs4==4.12.*
jinja2==3.1.*
aiofiles==24.1.*
httpx==0.28.*
aiosmtplib==3.0.*
